// src/entities/content/model/types.ts
export type ContentType =
    | 'html'
    | 'md'
    | 'csv'
    | 'txt'
    | 'xlsx'
    | 'xls'
    | 'docx'
    | 'pptx'
    | 'doc'
    | 'ppt'
    | 'image'
    | 'pdf'
    | 'video'
    | 'audio'
    | 'office'
    | 'recovery'
    | 'qr';

export type ToolType =
    | 'create'
    | 'recovery'
    | 'image'
    | 'pdf'
    | 'video'
    | 'audio'
    | 'office'
    | 'qr';